requirements.txt
arabic_reshaper==3.0.0
fitz==0.0.1.dev2
gradio==5.16.1
numpy==1.23.5
Pillow==11.1.0
tqdm==4.66.1
ultralytics==8.1.8
pymupdf== 1.25.4
